refactor(apex-lsp-compliant-services): extract worker runtimeContext module - W-24145209
Move the mutable process singletons and platform DI shims (role state,
log level, coordinator-assistance transport, tracing hooks, worker id,
resource-loader layer factory, remote-stdlib warmup) out of the
8.3k-line worker.platform.shared.ts into a new src/worker/runtimeContext.ts.
The new module sits at the bottom of the worker dependency graph so
later extractions can import it without creating cycles.

worker.platform.shared.ts imports the singletons for local use and
re-exports the previously-public surface, so both platform shells
(worker.platform.ts / .web.ts) and the tests import unchanged.

Also consolidates the two duplicate @salesforce/apex-lsp-shared imports
that eslint import/no-duplicates flagged once the edit busted a stale
wireit lint cache.

First slice of the worker.platform.shared.ts modularization (W-24145209).

W-24145209
